Added check_db.js file to list database content
The script can be run from a terminal inside the backend folder with the
command `node tests/check_db.js` and will display all of the data in the
database. The console.table() output for the recipes table is kinda
broken, but it's still meaningful and legible if you try hard enough.