Fixed bug with recipe cards on the Explore page
The recipe cards on the explore page used to not display the correct
number of minutes that a recipe needs. The cards on the User/Bookmarks
page did display them correctly, and were also colored. I changed the
span JSX elements in Search.jsx to match those in User.jsx, which fixed
the bug and made the elements more visually appealing.

I also made the View Recipe links stick to the bottom of their cards
regardless of the cards' width and the length of the description of
other cards on the same row.